# Speed up calculation of absorption probabilities