**IMPORTANT: Please search among the [Pull request](../pulls) before creating one.**

# Speed up calculation of absorption probabilities
<!--- Provide a general summary of your changes in the Title above -->

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] New feature (non-breaking change which adds functionality)

## Description
<!--- Clearly and concisely describe your changes. -->

Changes the computation of absorption probabilities according to @Marius1311 suggestion in #619.

## How has this been tested?
<!--- Describe in detail how you've tested your changes. -->

Unit tests already exist. `tests/test_cflare.py::TestCFLARE.test_abs_probs_negative` needs to be updated.

## Closes
<!--- Type `closes #XXXX` in your comment to auto-close the issue that your PR fixes (if such). -->
Closes #619.