@WeilerP 2021-06-28
@michalk8, `tests/test_cflare.py::TestCFLARE.test_abs_probs_negative` [here](https://github.com/theislab/cellrank/blob/f90ae2b2e87acf5d4f39492a803d060f8a688524/tests/test_cflare.py#L617) fails now since `cellrank.tl.estimators._base_estimator._solve_lin_system` now already returns the summed probabilities. Consequently, two values are now negative. I suggest to simply update the matching string accordingly. Or do you prefer a different solution?