@michalk8 2021-06-28
> I suggest to simply update the matching string accordingly. 

Agree, thanks!