@Marius1311 2021-06-28
Fantastic, thanks. Can we compare this in a unit test to previous, summed absorption probabilities? i.e. does the test you mention above do that on a test case problem @WeilerP?