@WeilerP 2021-06-28
> Fantastic, thanks. Can we compare this in a unit test to previous, summed absorption probabilities? i.e. does the test you mention above do that on a test case problem @WeilerP?

Unit tests checking the correct calculation of absorption probabilities already exist (e.g. [here](https://github.com/theislab/cellrank/blob/f4a08e62f9edb4d3e97f82c6129ad083d43cfd04/tests/test_gpcca.py#L443)) and are still passing, i.e. the old and new implementations give the same result. The test I mentioned is only designed to trigger `ValueError` which circumvents the actual calculation of absorption probabilities to speed up the unit test.