@Marius1311 2021-06-28
We just need some statement on the speedup now - that's why we did this in the first place. I would like to know whether we really get something in the range of 4x speedup on a medium sized dataset (~50k cells)