@WeilerP 2021-06-29
> I'll also check this out today on the reprogramming data to see what speedup we get.

Cool. I decided to test it on the arabidopsis dataset (which is similar in size I guess).