@Marius1311 2021-06-29
> > I'll also check this out today on the reprogramming data to see what speedup we get.
> 
> Nice! I decided to test it on the arabidopsis dataset (which is similar in size I guess).

yes! That's a good idea. Esp. for many terminal states, we should see significant speedup. If not, we have to check whether parallelization still does what it should. Original parallelization, @michalk8, was optimized to solve many linear solved, i.e. in the range of 60-150. With the new implementations, we will be looking at 2-10 linear solves, in most problems.