@Marius1311 2021-06-29
I tested this on a subsampled version of the reprogramming data (41.5k cells) on my laptop for 4 terminal states by calling
`g.compute_absorption_probabilities(n_jobs=n_jobs, solver='gmres', use_petsc=True)`. I get the following results

- old runtime (8 cores): 8m 42sec
- new runtime (8 cores): 1m 29sec
- new runtime (4 cores): 1m 25sec
- new runtime (1 core): 1m 15sec

So, this is great. I take two things from this:
1. It worked! No matter how many cores I use, I get at least 5x speedup. 
2. For such a low number of linear systems, parallelisation actually does not help and it's faster using a single core. This probably changes when we consider more terminal states.

@WeilerP, from my point of view, this can be merged as soon as tests pass. What sort of speedup did you see in your tests? Similar?