@Marius1311 2021-06-29
I should note one thing here: the reason why absorption probabilities take relatively long here is because I used the WOT kernel to compute the transition matrix. Unfortunately, solving an entropically-regularized OT problem leads to dense solutions, hence, the transition matrix has dense blocks on the 1st off-diagonal. Luckily, these are not that large because the number of cells per time point is reasonable. This becomes problematic for datasets with extremely large cell numbers distributed across few time-points - that would mean very large dense matrices which would massively slow down our linear solver. 

So, if I repeated the same experiment using e.g. the `VelocityKernel`, I would expect to see computation times much lower. In the CellRank paper, we compute absorption probs. for 100k cells on the server in 2 min (using the `VelocityKernel`).