@WeilerP 2021-06-29
@Marius1311, sorry for the delayed reply. I ran it on the [mouse gastrulation data](https://www.nature.com/articles/s41586-019-0933-9) containing ~90000 cells and 6 terminal states:

* old implementation (6 cores): 3:35min
* new impelementation (6 cores): 9sec
* new implementation (1 core): 28sec

So the new implementation is whopping 24 times (ish) faster.

> 2. For such a low number of linear systems, parallelisation actually does not help and it's faster using a single core. This probably changes when we consider more terminal states.

Just as a side remark: I would not expect an increase in performance using more than 4 cores since there were only 4 terminal states, i.e. only 4 processes to parallelize.