@Marius1311 2021-06-30
Thanks for these results! This is really great, it's a fantastic speedup. Being able to process 90k cells in 9 seconds is just fantastic!

As to the number of cores, you are right, for 4 terminal states, it doesn't make sense to use more than 4 cores. I was doing this to see whether using more cores than terminal states, in which case we should prevent the user from doing this by setting the number of cores equal to the number of terminal states and issuing a warning. I ended up not doing it because the performance loss wasn't great, so I thought it doesn't really matter. Feel free to open a PR to fix this - it's minor but user-friendly.