@michalk8 2021-06-30
>Just as a side remark: I would not expect an increase in performance using more than 4 cores since there were only 4 terminal states, i.e. only 4 processes to parallelize.
> As to the number of cores, you are right, for 4 terminal states, it doesn't make sense to use more than 4 cores. 

That's because n the parallelize function, you can't use more cores than jobs, I clip it.